requirements.txt
Flask==2.3.2
pymongo==4.4.1
python-dotenv==1.0.0
pydantic==2.0.3
PyJWT==2.8.0
pytest==7.3.1
pytest-cov==4.1.0
Flask-Cors==3.0.10
Flask-Caching==1.10.1
Flask-RESTful==0.3.9
Flask-Testing==0.8.1
pytest-flask==1.2.0
pytest-mock==3.6.1